Re-Style JSON Coder Website
Fixes #10.

I went with the nuclear option. Delete most of the CSS and revert to
vanilla Bootstrap, pulled from the CDN. Then, added back a relatively
small amount of CSS into main.css. The entire HTML/CSS architecture is
much simpler now. The site also has a clean, simple design, and should
be easier to work with. This style is simple enough that it should also
be simple to make small improvements to the style over time.

Importantly, the page does not need so scroll on a normal desktop-sized
browser window.